move gravity and hydrodynamics out of physics to top level folder
