+++ src/domain/mod.rs
@@ -18,10 +18,10 @@ use crate::communication::WorldSize;
 use crate::components::Mass;
 use crate::components::Position;
 use crate::components::Velocity;
+use crate::gravity;
+use crate::gravity::LeafData;
+use crate::gravity::MassMoments;
 use crate::named::Named;
-use crate::physics::gravity;
-use crate::physics::gravity::LeafData;
-use crate::physics::MassMoments;
 use crate::prelude::Particles;
 use crate::quadtree::QuadTreeConfig;
 use crate::quadtree::QuadTreeIndex;
