+++ src/gravity/tests/parallel.rs
@@ -13,10 +13,10 @@ use crate::components;
 use crate::components::Position;
 use crate::components::Velocity;
 use crate::domain::DomainDecompositionPlugin;
-use crate::physics::gravity::plugin::GravityPlugin;
-use crate::physics::gravity::GravityParameters;
-use crate::physics::gravity::LeafData;
-use crate::physics::gravity::Solver;
+use crate::gravity::plugin::GravityPlugin;
+use crate::gravity::GravityParameters;
+use crate::gravity::LeafData;
+use crate::gravity::Solver;
 use crate::physics::PhysicsPlugin;
 use crate::physics::Timestep;
 use crate::prelude::LocalParticle;
