+++ src/parameters.rs
@@ -1,10 +1,10 @@
 pub use crate::domain::DomainTreeParameters;
+pub use crate::gravity::GravityParameters;
+pub use crate::hydrodynamics::HydrodynamicsParameters;
+pub use crate::hydrodynamics::InitialGasEnergy;
 pub use crate::io::input::InputParameters;
 pub use crate::io::output::parameters::*;
 pub use crate::performance_parameters::PerformanceParameters;
-pub use crate::physics::gravity::GravityParameters;
-pub use crate::physics::hydrodynamics::HydrodynamicsParameters;
-pub use crate::physics::hydrodynamics::InitialGasEnergy;
 pub use crate::physics::SimulationParameters;
 pub use crate::quadtree::QuadTreeConfig;
 pub use crate::visualization::VisualizationParameters;
