+++ src/visualization/mod.rs
@@ -26,8 +26,8 @@ use crate::communication::CommunicationType;
 use crate::communication::Rank;
 use crate::components::Position;
 use crate::domain::determine_global_extent_system;
+use crate::gravity;
 use crate::named::Named;
-use crate::physics::gravity;
 use crate::physics::StopSimulationEvent;
 use crate::prelude::LocalParticle;
 use crate::quadtree::QuadTreeVisualizationPlugin;
